Modified the darkening script
